Update SparkFun_MAG3110.cpp
Fixes #1, according to datasheet section 5.1.2 "X-axis, Y-axis, and Z-axis 16-bit output sample data of the magnetic field strength expressed as signed 2's complement numbers."

+++ src/SparkFun_MAG3110.cpp
@@ -99,9 +99,9 @@ void MAG3110::readMag(int* x, int* y, int* z){
 	}
 
 	// Put data into referenced variables
-	*x = (int) values[0];
-	*y = (int) values[1];
-	*z = (int) values[2];
+	*x = (int16_t) values[0];
+	*y = (int16_t) values[1];
+	*z = (int16_t) values[2];
 
 }
